// src/app/asset/rejected-assets/rejected-assets.interface.ts
export interface RejectedAssetsInterface {
  id: string;
  assetId: string;
  images: string[];
  assetName: string;
  category: string;
  sellerName: string;
  submissionDate: string;
  rejectedDate: string;
  adminRemarks: string;
  rejectedCount: number;
  key: string;
  coverImage: string;
  updatedAt: string;
}